Fix/fee fix

+++ services/JimService.js
@@ -2,38 +2,26 @@ import Message from '../models/message';
 import { gravity } from '../config/gravity';
 import metis from '../config/metis';
 import { loadInitialJFSImage } from '../utils/utils';
-
+import {FeeManager, feeManagerSingleton} from "./FeeManager";
+import jupiterApiService from "./jupiterAPIService";
+import fs from "fs";
 const FormData = require('form-data');
 const axios = require('axios');
 const { getPNTokensAndSendPushNotification, errorMessageHandler } = require('./messageService');
-const defaultFileBase64Encoded = loadInitialJFSImage();
-const defaultFileName = 'pixi.jpg';
-const delayUntilAccountIsCreated = 30000;
-
-const createStorageAndLoadImage = async (account, passphrase, encryptionPassword) => {
-  setTimeout(async () => {
-      console.log('[createStorageAndLoadImage]: Async Start');
-      console.log('Pixi.jpg ', defaultFileBase64Encoded);
-      if (!account || !passphrase || !encryptionPassword) {
-        console.log('Missing parameters required.', account, passphrase, encryptionPassword);
-        throw new Error('Missing parameters required.');
-      }
-      jupiterUpload(account, passphrase, encryptionPassword, defaultFileBase64Encoded, defaultFileName)
-          .then((response) => {
-            const { url } = response.data;
-            console.log('your storage has been created: ', account, url);
-            console.log('[createStorageAndLoadImage]: Async End');
-          })
-          .catch((error) => {
-            console.log('Something went wrong creatng your storage: ', account, error);
-            console.log('[createStorageAndLoadImage]: Async Error');
-            throw new Error('Something went wrong creatng your storage: ', account, error);
-          });
-    }, delayUntilAccountIsCreated);
-}
+const accountPropertyFee = feeManagerSingleton.getFee(FeeManager.feeTypes.account_property);  //ACCOUNT_PROPERTY_FEE
+const logger = require('../utils/logger')(module);
 
 
-const jupiterUpload = (
+/**
+ *
+ * @param jupAccount
+ * @param passphraseAccount
+ * @param passwordAccount
+ * @param fileBase64Encoded
+ * @param fileName
+ * @returns {Promise<AxiosResponse>}
+ */
+const UploadAnImageButFirstCheckForStorageAndCreateIfMissing = (
   jupAccount,
   passphraseAccount,
   passwordAccount,
@@ -41,7 +29,7 @@ const jupiterUpload = (
   fileName,
 ) => {
 
-  if ( !jupAccount || !passphraseAccount || !passwordAccount ) {
+  if ( !jupAccount || !passphraseAccount || !passwordAccount || !fileBase64Encoded ) {
     throw new Error('Missing params, Account, passphrase ans password are required');
   }
 
@@ -57,6 +45,8 @@ const jupiterUpload = (
     },
   };
 
+  // @TODO  In the near future /api/v1/storage will return a jobID so we need to poll for the job.
+  // @TODO this code (post:storage) might not be needed anymore since a metis signup will create the storage.
   return axios.post(`${process.env.JIM_SERVER}/api/v1/signin`, dataLogin)
     .then((response) => {
       defaultHeader.headers.Authorization = `Bearer ${response.data.token}`;
@@ -66,38 +56,32 @@ const jupiterUpload = (
       if (!responseStorageInfo.data) {
         return axios.post(`${process.env.JIM_SERVER}/api/v1/storage`, {}, defaultHeader);
       }
-
       return responseStorageInfo.data;
     })
     .then(() => {
       const buffer = Buffer.from(fileBase64Encoded, 'base64');
       const form = new FormData();
       form.append('file', buffer, fileName);
-
       defaultHeader.headers['Content-Type'] = `multipart/form-data; boundary=${form.getBoundary()}`;
       return axios.post(`${process.env.JIM_SERVER}/api/v1/file`, form, defaultHeader);
     });
 };
 
-module.exports = {
-  channelProfileDelete: (req, res) => {
-    const { channel } = req;
-    if (!channel) {
-      return res.status(500).json({ msg: 'No channel info' });
-    }
+const profileDelete = (passphrase, account, password) => {
 
-    const { passphrase, account, password } = channel.channel_record;
-    const dataLogin = { account, passphrase, password };
-    const defaultHeader = { headers: { Authorization: '' } };
-    const addressBreakdown = account.split('-');
+  const dataLogin = { account, passphrase, password };
+  const defaultHeader = { headers: { Authorization: '' } };
+  const addressBreakdown = account.split('-');
 
+  if(!(addressBreakdown.length === 5) ){
+    throw new Error('bla blab')
+  }
 
-    axios.post(`${process.env.JIM_SERVER}/api/v1/signin`, dataLogin)
+  return axios.post(`${process.env.JIM_SERVER}/api/v1/signin`, dataLogin)
       .then((response) => {
         defaultHeader.headers.Authorization = `Bearer ${response.data.token}`;
-        return axios.get(`${process.env.JIM_SERVER}/api/v1/storage`, defaultHeader);
+        return gravity.getAccountProperties({ recipient: channel.channel_record.account })
       })
-      .then(() => gravity.getAccountProperties({ recipient: channel.channel_record.account }))
       .then((userProperties) => {
         const profilePicture = userProperties.properties.find(property => property.property.includes('profile_picture'));
 
@@ -106,8 +90,13 @@ module.exports = {
         }
         return profilePicture.value;
       })
-      .then((url) => {
-        const id = url.split('/').pop();
+      .then((profilePicUrl) => {
+        //@TODO cannt assume the url is what you are expecting!
+        const urlArray = profilePicUrl.split('/');
+        if(urlArray.length == 0){
+          throw new Error('Theres a problem with the jim image url');
+        }
+        const id = urlArray.pop();
         return axios.delete(`${process.env.JIM_SERVER}/api/v1/file/${id}`, defaultHeader);
       })
       .then(() => {
@@ -115,17 +104,108 @@ module.exports = {
           passphrase,
           recipient: account,
           value: '',
-          feeNQT: process.env.PROPERTY_FEE,
+          feeNQT: accountPropertyFee,
           property: `profile_picture-${addressBreakdown[addressBreakdown.length - 1]}`,
         };
         return gravity.setAcountProperty(accountPropertyParams);
       })
-      .then(() => res.status(200).json({ url: '' }))
-      .catch((error) => {
-        console.log('Something went wrong', error);
-        res.status(500).json(errorMessageHandler(error));
-      });
+}
+
+
+
+
+
+module.exports = {
+
+  uploadPixiImageAndWait: async (account, passphrase, encryptionPassword) => {
+    logger.debug('uploadPixiImageAndWait()');
+    if (!account || !passphrase || !encryptionPassword) {
+      logger.debug('Missing parameters required.', account, passphrase, encryptionPassword);
+      throw new Error('Missing parameters required.');
+    }
+    const fileName = 'pixi.jpg';
+    const initialPixiImage = fs.readFileSync('./pixi.jpg', {encoding: 'base64'});
+    return UploadAnImageButFirstCheckForStorageAndCreateIfMissing(account, passphrase, encryptionPassword, initialPixiImage, fileName);
   },
+  // userProfileDelete: (req, res) => {
+  //   const { user } = req;
+  //   if (!user) {
+  //     return res.status(500).json({ msg: 'No user info' });
+  //   }
+  //   const userAccount = JSON.parse(gravity.decrypt(user.accountData));
+  //   const { passphrase, account, password } = userAccount;
+  //   const dataLogin = { account, passphrase, password };
+  //   const defaultHeader = { headers: { Authorization: '' } };
+  //   const addressBreakdown = account.split('-');
+  //   axios.post(`${process.env.JIM_SERVER}/api/v1/signin`, dataLogin)
+  //       .then((response) => {
+  //         defaultHeader.headers.Authorization = `Bearer ${response.data.token}`;
+  //         return axios.get(`${process.env.JIM_SERVER}/api/v1/storage`, defaultHeader);
+  //       })
+  //       .then(() => gravity.getAccountProperties({ recipient: userAccount.account }))
+  //       .then((userProperties) => {
+  //         const profilePicture = userProperties.properties.find(property => property.property.includes('profile_picture'));
+  //
+  //         if (!profilePicture || !profilePicture.value) {
+  //           throw new Error('The account does not have profile picture');
+  //         }
+  //         return profilePicture.value;
+  //       })
+  //       .then((url) => {
+  //         const id = url.split('/').pop();
+  //         return axios.delete(`${process.env.JIM_SERVER}/api/v1/file/${id}`, defaultHeader);
+  //       })
+  //       .then(() => {
+  //         const accountPropertyParams = {
+  //           passphrase,
+  //           recipient: account,
+  //           value: '',
+  //           feeNQT: accountPropertyFee,
+  //           property: `profile_picture-${addressBreakdown[addressBreakdown.length - 1]}`,
+  //         };
+  //         return gravity.setAcountProperty(accountPropertyParams);
+  //       })
+  //       .then(() => res.status(200).json({ url: '' }))
+  //       .catch((error) => {
+  //         logger.debug('Something went wrong', error);
+  //         res.status(500).json(errorMessageHandler(error));
+  //       });
+  // },
+
+  deleteUserProfile: (req, res) => {
+    const { user } = req;
+    if (!user) {
+      return res.status(500).json({ msg: 'No user info' });
+    }
+    const userAccount = JSON.parse(gravity.decrypt(user.accountData));
+    const { passphrase, account, password } = userAccount;
+
+    profileDelete(passphrase, account, password)
+        .then(() => res.status(200).json({ url: '' }))
+        .catch((error) => {
+          logger.debug('Something went wrong', error);
+          res.status(500).json(errorMessageHandler(error));
+        });
+  },
+
+  deleteChannelProfile: (req, res) => {
+    const { channel } = req;
+
+    if (!channel) {
+      return res.status(500).json({ msg: 'No channel info' });
+    }
+    const { passphrase, account, password } = channel.channel_record;
+
+    profileDelete(passphrase, account, password)
+        .then(() => res.status(200).json({ url: '' }))
+        .catch((error) => {
+          logger.debug('Something went wrong', error);
+          res.status(500).json(errorMessageHandler(error));
+        });
+  },
+
+
+  //@TODO pass in channel not req
   channelProfileDisplay: (req, res) => {
     const { channel } = req;
 
@@ -141,12 +221,12 @@ module.exports = {
         res.send(response);
       })
       .catch((error) => {
-        console.log('Something went wrong', error);
+        logger.debug('Something went wrong', error);
         res.status(500).json(errorMessageHandler(error));
       });
   },
   channelProfileUpload: (req, res) => {
-    console.log('[channelProfileUpload]: Start');
+    logger.debug('[channelProfileUpload]: Start');
     const { user, channel } = req;
     const { passphrase, account, password } = channel.channel_record;
     const fileBase64Encoded = req.body.base64Image;
@@ -157,43 +237,40 @@ module.exports = {
       return res.status(400).json({ msg: 'Missing parameters required.' });
     }
 
-    jupiterUpload(account, passphrase, password, fileBase64Encoded, fileName)
+    UploadAnImageButFirstCheckForStorageAndCreateIfMissing(account, passphrase, password, fileBase64Encoded, fileName)
       .then((response) => {
-        console.log(`jupiterUpload().then()`)
-        console.log(response)
+        logger.debug(`jupiterUpload().then()`)
+        logger.debug(response)
         const { url } = response.data;
         const accountPropertyParams = {
           passphrase,
           recipient: account,
           value: url,
-          feeNQT: process.env.PROPERTY_FEE,
+          feeNQT: accountPropertyFee,
           property: `profile_picture-${addressBreakdown[addressBreakdown.length - 1]}`,
         };
 
-        console.log(accountPropertyParams);
+        logger.debug(accountPropertyParams);
 
         return Promise.all([url, gravity.setAcountProperty(accountPropertyParams)]);
       })
       .then((response) => {
-        console.log(`jupiterUpload().then().then()`)
-        console.log(response)
-
-
+        logger.debug(`jupiterUpload().then().then()`)
+        // logger.debug(response)
         const [url, accountPropertyResponse] = response;
         if (accountPropertyResponse && accountPropertyResponse.errorDescription) {
           throw new Error(accountPropertyResponse.errorDescription);
         }
-
         res.status(200).json({ url });
       })
       .catch((error) => {
-        console.log('Something went wrong', error.response);
+        logger.debug('Something went wrong', error);
         const response = errorMessageHandler(error);
         res.status(400).send(response);
       });
   },
   userProfileUpload: (req, res) => {
-    console.log('[userProfileUpload]: Start');
+    logger.debug('[userProfileUpload]: Start');
     const { user } = req;
     if (!user) {
       return res.status(500).json({ msg: 'No user info' });
@@ -202,20 +279,24 @@ module.exports = {
     const { passphrase, account, encryptionPassword } = userAccount;
     const fileBase64Encoded = req.body.base64Image;
     const fileName = userAccount.account;
-    const addressBreakdown = account.split('-');
+    const addressBreakdown = account.split('-');  // ex JUP-D5A4-2LSA-8VLJ-6MGJM
+
+    if(addressBreakdown.length !== 5 ){
+      return res.status(500).json({msg: 'Internal Error'});
+    }
 
     if (!(fileBase64Encoded) || !user || !fileName) {
       return res.status(400).json({ msg: 'Missing parameters required.' });
     }
 
-    jupiterUpload(account, passphrase, encryptionPassword, fileBase64Encoded, fileName)
+    UploadAnImageButFirstCheckForStorageAndCreateIfMissing(account, passphrase, encryptionPassword, fileBase64Encoded, fileName)
         .then((response) => {
           const { url } = response.data;
           const accountPropertyParams = {
             passphrase,
             recipient: account,
             value: url,
-            feeNQT: process.env.PROPERTY_FEE,
+            feeNQT: accountPropertyFee,
             property: `profile_picture-${addressBreakdown[addressBreakdown.length - 1]}`,
           };
           return Promise.all([url, gravity.setAcountProperty(accountPropertyParams)]);
@@ -225,15 +306,15 @@ module.exports = {
           if (accountPropertyResponse && accountPropertyResponse.errorDescription) {
             throw new Error(accountPropertyResponse.errorDescription);
           }
-          console.log('[jupiterUpload]:', url);
+          logger.debug('[jupiterUpload]:', url);
           res.status(200).json({ url });
         })
         .catch((error) => {
-          console.log('Something went wrong', error);
+          logger.debug('Something went wrong', error);
           res.status(500).json(errorMessageHandler(error));
         });
   },
-  userJimLogin: (req, res) =>{
+  jimSignin: (req, res) =>{
     const { user } = req;
     if (!user) {
       return res.status(500).json({ msg: 'No user info' });
@@ -251,52 +332,8 @@ module.exports = {
         return res.status(500).json({msg: 'Error generating token'});
       })
       .catch((error) => {
-        console.log('Something went wrong whit JIM login', error);
+        logger.debug('Something went wrong whit JIM login', error);
         res.status(500).json({ msg: 'Something went wrong whit JIM login', error });
-      });
-  },
-  userProfileDelete: (req, res) => {
-    const { user } = req;
-    if (!user) {
-      return res.status(500).json({ msg: 'No user info' });
-    }
-    const userAccount = JSON.parse(gravity.decrypt(user.accountData));
-    const { passphrase, account, password } = userAccount;
-    const dataLogin = { account, passphrase, password };
-    const defaultHeader = { headers: { Authorization: '' } };
-    const addressBreakdown = account.split('-');
-    axios.post(`${process.env.JIM_SERVER}/api/v1/signin`, dataLogin)
-      .then((response) => {
-        defaultHeader.headers.Authorization = `Bearer ${response.data.token}`;
-        return axios.get(`${process.env.JIM_SERVER}/api/v1/storage`, defaultHeader);
-      })
-      .then(() => gravity.getAccountProperties({ recipient: userAccount.account }))
-      .then((userProperties) => {
-        const profilePicture = userProperties.properties.find(property => property.property.includes('profile_picture'));
-
-        if (!profilePicture || !profilePicture.value) {
-          throw new Error('The account does not have profile picture');
-        }
-        return profilePicture.value;
-      })
-      .then((url) => {
-        const id = url.split('/').pop();
-        return axios.delete(`${process.env.JIM_SERVER}/api/v1/file/${id}`, defaultHeader);
-      })
-      .then(() => {
-        const accountPropertyParams = {
-          passphrase,
-          recipient: account,
-          value: '',
-          feeNQT: process.env.PROPERTY_FEE,
-          property: `profile_picture-${addressBreakdown[addressBreakdown.length - 1]}`,
-        };
-        return gravity.setAcountProperty(accountPropertyParams);
-      })
-      .then(() => res.status(200).json({ url: '' }))
-      .catch((error) => {
-        console.log('Something went wrong', error);
-        res.status(500).json(errorMessageHandler(error));
       });
   },
   userProfileDisplay: (req, res) => {
@@ -314,12 +351,12 @@ module.exports = {
         res.send(response);
       })
       .catch((error) => {
-        console.log('Something went wrong', error);
+        logger.debug('Something went wrong', error);
         res.status(500).json(errorMessageHandler(error));
       });
   },
   fileUpload: (req, res) => {
-    console.log('[fileUpload]: Start');
+    logger.debug('[fileUpload]: Start');
     const { user, channel } = req;
     const fileBase64Encoded = req.body.file.data;
     const fileName = req.body.file.name;
@@ -329,7 +366,7 @@ module.exports = {
       return res.status(400).json({ msg: 'Missing parameters required.' });
     }
 
-    jupiterUpload(
+    UploadAnImageButFirstCheckForStorageAndCreateIfMissing(
       channel.channel_record.account,
       channel.channel_record.passphrase,
       channel.channel_record.password,
@@ -351,7 +388,7 @@ module.exports = {
         const tableData = channel.channel_record;
         const userData = JSON.parse(gravity.decrypt(user.accountData));
         const messageModel = new Message(dataMessage);
-        return messageModel.sendMessage(userData, tableData, messageModel.record);
+        return messageModel.sendRecord(userData, tableData, messageModel.record);
       })
       .then(() => metis.getMember({
         channel: channel.channel_record.account,
@@ -371,19 +408,19 @@ module.exports = {
       })
       .then(() => res.status(200).json({}))
       .catch((error) => {
-        console.log('Something went wrong', error);
+        logger.debug('Something went wrong', error);
         res.status(500).json(errorMessageHandler(error));
       });
   },
 
-  userStorageCreate: (account, passphrase, encryptionPassword) => {
-    console.log('[userStorageCreate]: Start');
-    createStorageAndLoadImage(account, passphrase, encryptionPassword);
-    console.log('[userStorageCreate]: Creating...');
-  },
-  channelStorageCreate: async (account, passphrase, encryptionPassword) => {
-    console.log('[channelStorageCreate]: Start');
-    createStorageAndLoadImage(account, passphrase, encryptionPassword);
-    console.log('[channelStorageCreate]: Creating...');
-  },
+  // userStorageCreate: async (account, passphrase, encryptionPassword) => {
+  //   logger.debug('[userStorageCreate]: Start');
+  //   await uploadPixiImageAndWait(account, passphrase, encryptionPassword);
+  //   logger.debug('[userStorageCreate]: Creating...');
+  // },
+  // channelStorageCreate: async (account, passphrase, encryptionPassword) => {
+  //   logger.debug('[channelStorageCreate]: Start');
+  //   await uploadPixiImageAndWait(account, passphrase, encryptionPassword);
+  //   logger.debug('[channelStorageCreate]: Creating...');
+  // },
 };
